[#67] feat: User 도메인 객체 생성
API 요청을 위한 Controller를 분리하였습니다. (현재는 데모를 위해서 요청경로를 기존과 같게 유지하였습니다.)
User 목록을 잘 불러올 수 있는지, User를 잘 생성할 수 있는지의 여부를 확인할 수 있습니다. (추후 테스트코드로 이동)
